requirements.txt
capstone == 4.0.2
coverage == 5.3
portion == 2.1.3
sortedcontainers == 2.2.2
unicorn == 1.0.2
pyelftools == 0.27
